copilot/llamacpp/aws-copilot-inference-service/requirements.txt
aiohttp==3.7.4
gunicorn==20.1.0
llama-cpp-python==0.1.78
